+++ CHANGELOG.md
@@ -33,6 +33,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Parse INFO.MATEID as `mate_id`, when available (#28)
 - Parse AZLENGTH and AZQUAL values, when available (#31)
 - Parse fusion variants INFO field (#32)
+- Set fusion genes from fusion variant INFO (#33)
 ### Fixed
 - Normalize chromosome names by stripping the chr prefix (#4)
 - Parsing of end chromosome (#29)